✨ feat: /api/letter/reception/:letterId 에 400 에러 적용
