+++ src/pages/LetterReceptionPage/hooks/useLetterTag.ts
@@ -2,15 +2,19 @@ import { useQuery } from '@tanstack/react-query';
 import letterOptions from '@/api/letter/queryOptions';
 import { WORRY_DICT } from '@/constants/users';
 import { Reception } from '@/types/letter';
-
 interface ReceptionLetterType extends Reception {
   tagList: string[];
 }
 
-const useLetterTag = (letterId: string) => {
-  const { data } = useQuery({
-    ...letterOptions.singleReception(Number(letterId)),
+const useLetterTag = (letterId: number) => {
+  const { data, error } = useQuery({
+    ...letterOptions.singleReception(letterId),
   });
+
+  if (error) {
+    /** 에러 나면 ???  */
+    console.error(error.message);
+  }
 
   /**
    * TODO: 나이, 성별 값 배열에 넣기
