+++ src/pages/LetterReceptionPage/intex.tsx
@@ -1,11 +1,11 @@
 import { useNavigate, useParams } from 'react-router-dom';
-import { css } from '@emotion/react';
-import createFunnel from '@/components/Funnel/createFunnel';
 import { ROUTER_PATHS } from '@/router';
+import createFunnel from '@/components/Funnel/createFunnel';
 import useLetterTag from './hooks/useLetterTag';
 import ReceivedLetter from './ReceivedLetter';
 import ReplyToLetter from './ReplyToLetter';
 import ReceptionHeader from './components/ReceptionHeader';
+import style from './styles';
 
 const { Funnel, Step, useFunnel } = createFunnel([
   'ReceivedLetter',
@@ -13,16 +13,12 @@ const { Funnel, Step, useFunnel } = createFunnel([
 ] as const);
 
 const LetterReceptionPage = () => {
+  const navigate = useNavigate();
   const { letterId } = useParams();
 
-  /**
-   * letterId 가 유효하지 않을때에는?
-   * letterId 가 숫자가 아닌 값일 땐 "존재하지 않는 페이지 입니다"? 보여주기?
-   */
-  const { letter } = useLetterTag(letterId || '1');
+  const { letter } = useLetterTag(Number(letterId));
 
   const { step, setStep, toPrev } = useFunnel();
-  const navigate = useNavigate();
 
   return (
     <div css={style.container}>
@@ -47,15 +43,6 @@ const LetterReceptionPage = () => {
       </Funnel>
     </div>
   );
-};
-
-const style = {
-  container: css`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  `,
 };
 
 export default LetterReceptionPage;
