+++ src/pages/LetterReceptionPage/styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@emotion/react';
+
+const style = {
+  container: css`
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    height: 100%;
+  `,
+};
+
+export default style;
